Shipping a packet
How much ever care is taken, there are many chances that we may forget
something while we are travelling. What if we forget something that
affects our purpose of travel which is very important? We will not be
always left with an option to go back and get it. One solution for this
is approaching a shipping company and get the item shipped, but if in
case you need them to be shipped over night then the amount that the
shipping company will charge is pretty much high. In such scenarios the
option that may workout will be asking someone who is travelling to your
place, to get it. A question that arises here is "How to find that one
particular person who is travelling to the place I am at?" This project
is all about providing a platform to the one who is in need of a carrier
who is ready to carry the packet with him/her.  This application brings
together the people who want to send or receive a packet and the people
who are ready to carry that packet.